--- .env
@@ -1,3 +1,0 @@
-MODEL="gpt-4o-2024-11-20"
-MAX_ITERATIONS=3
-DESIRED_COVERAGE=70
