+++ tests_integration/run_test_all.py
@@ -2,7 +2,6 @@ import argparse
 import os
 import sys
 
-from enum import Enum
 from typing import Any, Iterable
 
 import docker
@@ -11,17 +10,13 @@ from dotenv import load_dotenv
 
 from cover_agent.CustomLogger import CustomLogger
 
+import tests_integration.constants as constants
 from run_test_with_docker import run_test
+from scenarios import TESTS
 
 
 load_dotenv()
 logger = CustomLogger.get_logger(__name__)
-
-
-class CoverageType(Enum):
-    LCOV = "lcov"
-    COBERTURA = "cobertura"
-    JACOCO = "jacoco"
 
 
 def stream_docker_logs(response: Iterable[bytes|dict[str, str]]) -> None:
@@ -45,7 +40,7 @@ def build_docker_image(client: docker.DockerClient, dockerfile: str, platform: s
     Force build for x86_64 architecture (`linux/amd64`) even for Apple Silicon currently.
     """
     try:
-        logger.info(f"Building image from {dockerfile}...")
+        logger.info(f"Building Docker image from {dockerfile}...")
         response = client.api.build(
             path=".",
             dockerfile=dockerfile,
@@ -55,7 +50,7 @@ def build_docker_image(client: docker.DockerClient, dockerfile: str, platform: s
         )
         stream_docker_logs(response)
     except DockerException as e:
-        logger.error(f"Error building image: {e}")
+        logger.error(f"Error building Docker image: {e}")
         exit(1)
 
 
@@ -66,7 +61,7 @@ def run_docker_container(
         command: str="/bin/sh -c 'tail -f /dev/null'",
 ) -> None:
     try:
-        logger.info(f"Running container for {image}...")
+        logger.info(f"Running Docker container for {image}...")
         container = client.containers.run(
             image=image,
             command=command,
@@ -83,19 +78,19 @@ def run_docker_container(
                 
         exit_code = container.wait()["StatusCode"]
         if exit_code != 0:
-            raise DockerException(f"Container exited with status {exit_code}")
+            raise DockerException(f"Docker container exited with status {exit_code}.")
             
     except DockerException as e:
-        logger.error(f"Error running container: {e}")
+        logger.error(f"Error running Docker container: {e}")
         if "container" in locals():
             container.remove(force=True)
         sys.exit(1)
 
 
 def main():
-    parser = argparse.ArgumentParser(description="Run all tests with Docker.")
-    parser.add_argument("--model", default=os.getenv("MODEL"), help="Model name.")
-    parser.add_argument("--run-installer", action="store_true", help="Run the installer within a Docker container.")
+    parser = argparse.ArgumentParser(description="Args for running tests with Docker.")
+    parser.add_argument("--model", default=constants.MODEL, help="Which LLM model to use.")
+    parser.add_argument("--run-installer", action="store_true", help="Build `cover-agent` binary.")
     args = parser.parse_args()
 
     model = args.model
@@ -103,6 +98,7 @@ def main():
     client = docker.from_env()
 
     if run_installer:
+        # Compile `cover-agent` binary
         build_docker_image(client, "Dockerfile")
         os.makedirs("dist", exist_ok=True)
         dist_path = os.path.join(os.getcwd(), "dist")
@@ -110,123 +106,16 @@ def main():
             client, "cover-agent-installer", {dist_path: {"bind": "/app/dist", "mode": "rw"}}, command="make installer"
         )
 
-    tests = [
-        # C Calculator Example
-        {
-            "docker_image": "embeddeddevops/c_cli:latest",
-            "source_file_path": "calc.c",
-            "test_file_path": "test_calc.c",
-            "code_coverage_report_path": "coverage_filtered.info",
-            "test_command": r"sh build_and_test_with_coverage.sh",
-            "coverage_type": CoverageType.LCOV.value,
-            "max_iterations": 4,
-            "desired_coverage": 50,
-        },
-        # C++ Calculator Example
-        {
-            "docker_image": "embeddeddevops/cpp_cli:latest",
-            "source_file_path": "calculator.cpp",
-            "test_file_path": "test_calculator.cpp",
-            "code_coverage_report_path": "coverage.xml",
-            "test_command": r"sh build_and_test_with_coverage.sh",
-            "coverage_type": CoverageType.COBERTURA.value,
-        },
-        # C# Calculator Web Service
-        # TODO: Fix test_command here as it fails in macOS
-        {
-            "docker_image": "embeddeddevops/csharp_webservice:latest",
-            "source_file_path": "CalculatorApi/CalculatorController.cs",
-            "test_file_path": "CalculatorApi.Tests/CalculatorControllerTests.cs",
-            "code_coverage_report_path": "CalculatorApi.Tests/TestResults/coverage.cobertura.xml",
-            "test_command": (
-                r'dotnet test --collect:"XPlat Code Coverage" CalculatorApi.Tests/ && find . '
-                r'-name "coverage.cobertura.xml" -exec mv {} CalculatorApi.Tests/TestResults/coverage.cobertura.xml \;'
-            ),
-            "coverage_type": CoverageType.COBERTURA.value,
-            "desired_coverage": "50",
-        },
-        # Go Webservice Example
-        {
-            "docker_image": "embeddeddevops/go_webservice:latest",
-            "source_file_path": "app.go",
-            "test_file_path": "app_test.go",
-            "test_command": (
-              r"go test -coverprofile=coverage.out && gocov convert coverage.out | gocov-xml > coverage.xml"
-            ),
-            "max_iterations": 4,
-        },
-        # Java Gradle example
-        # TODO: Figure out why it fails in macOS with `timed out`
-        {
-            "docker_image": "embeddeddevops/java_gradle:latest",
-            "source_file_path": "src/main/java/com/davidparry/cover/SimpleMathOperations.java",
-            "test_file_path": "src/test/groovy/com/davidparry/cover/SimpleMathOperationsSpec.groovy",
-            "test_command": r"./gradlew clean test jacocoTestReport",
-            "coverage_type": CoverageType.JACOCO.value,
-            "code_coverage_report_path": "build/reports/jacoco/test/jacocoTestReport.csv",
-        },
-        # Java Spring Calculator example
-        {
-            "docker_image": "embeddeddevops/java_spring_calculator:latest",
-            "source_file_path": "src/main/java/com/example/calculator/controller/CalculatorController.java",
-            "test_file_path": "src/test/java/com/example/calculator/controller/CalculatorControllerTest.java",
-            "test_command": r"mvn verify",
-            "coverage_type": CoverageType.JACOCO.value,
-            "code_coverage_report_path": "target/site/jacoco/jacoco.csv",
-        },
-        # VanillaJS Example
-        {
-            "docker_image": "embeddeddevops/js_vanilla:latest",
-            "source_file_path": "ui.js",
-            "test_file_path": "ui.test.js",
-            "test_command": r"npm run test:coverage",
-            "code_coverage_report_path": "coverage/coverage.xml",
-        },
-        # Python FastAPI Example
-        {
-            "docker_image": "embeddeddevops/python_fastapi:latest",
-            "source_file_path": "app.py",
-            "test_file_path": "test_app.py",
-            "test_command": r"pytest --cov=. --cov-report=xml --cov-report=term",
-            "model": "gpt-4o-mini",
-        },
-        # React Calculator Example
-        {
-            "docker_image": "embeddeddevops/react_calculator:latest",
-            "source_file_path": "src/modules/Calculator.js",
-            "test_file_path": "src/tests/Calculator.test.js",
-            "test_command": r"npm run test",
-            "code_coverage_report_path": "coverage/cobertura-coverage.xml",
-            "desired_coverage": "55",
-        },
-        # Ruby Sinatra Example
-        {
-            "docker_image": "embeddeddevops/ruby_sinatra:latest",
-            "source_file_path": "app.rb",
-            "test_file_path": "test_app.rb",
-            "test_command": r"ruby test_app.rb",
-            "code_coverage_report_path": "coverage/coverage.xml",
-        },
-        # TypeScript Calculator Example
-        {
-            "docker_image": "embeddeddevops/typescript_calculator:latest",
-            "source_file_path": "src/modules/Calculator.ts",
-            "test_file_path": "tests/Calculator.test.ts",
-            "test_command": r"npm run test",
-            "code_coverage_report_path": "coverage/cobertura-coverage.xml",
-        },
-    ]
-
-    for test in tests:
+    for test in TESTS:
         test_args = argparse.Namespace(
             docker_image=test["docker_image"],
             source_file_path=test["source_file_path"],
             test_file_path=test["test_file_path"],
             code_coverage_report_path=test.get("code_coverage_report_path", "coverage.xml"),
             test_command=test["test_command"],
-            coverage_type=test.get("coverage_type", CoverageType.COBERTURA.value),
-            max_iterations=test.get("max_iterations", os.getenv("MAX_ITERATIONS")),
-            desired_coverage=test.get("desired_coverage", os.getenv("DESIRED_COVERAGE")),
+            coverage_type=test.get("coverage_type", constants.CoverageType.COBERTURA.value),
+            max_iterations=test.get("max_iterations", constants.MAX_ITERATIONS),
+            desired_coverage=test.get("desired_coverage", constants.DESIRED_COVERAGE),
             model=test.get("model", model),
             api_base=os.getenv("API_BASE", ""),
             openai_api_key=os.getenv("OPENAI_API_KEY", ""),
